pyproject.toml
[project]
name = "bioinformatics-scout"
version = "0.1.0"
description = "Multi-agent literature discovery pipeline for bioinformatics research"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.0.7",
    "biopython>=1.84",
    "arxiv>=2.1.3",
    "gradio>=5.12.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
